Rename task-queue to queue
