+++ package.json
@@ -15,6 +15,6 @@
     "workspaces": [
         "backend",
         "frontend",
-        "task-queue"
+        "queue"
     ]
 }
