+++ queue/lib/parser.js
