+++ queue/lib/queue.js
