+++ queue/lib/queueRedis.js
