+++ queue/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "task-queue",
+  "name": "queue",
   "version": "0.0.1",
   "main": "index.js",
   "license": "MIT",
